Refactor authentication flow and enhance user experience
- Removed the Splash page and integrated direct login functionality into the main app.
- Implemented Google sign-in button for user authentication in the header.
- Updated ProtectedRoute logic to simplify access control.
- Adjusted map interaction to require authentication for adding new pins.
- Enhanced PinContent component to conditionally render comment functionality based on user authentication status.
- Updated smoke checklist to reflect changes in authentication flow.
